backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
cryptography==43.0.1
PyJWT==2.9.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0
ruff==0.6.4
black==24.8.0
pytest==8.3.2
